+++ AMD Power Gadget/ProcessorModel.swift
@@ -59,7 +59,7 @@ class ProcessorModel {
                                       &outputStr, &outputStrCount)
         AMDRyzenCPUPowerManagementVersion = outputStrCount > 0 ? String(cString: Array(outputStr[0...min(outputStrCount - 1, outputStr.count - 1)])) : ""
 
-        let compatVers = ["3.0.0", "3.1.0", "3.2.0", "3.3.0", "3.3.1", "3.4.0", "3.5.0", "3.6.0", "3.7.0", "3.8.0", "3.9.0", "3.10.0", "3.11.0", "3.12.0"]
+        let compatVers = ["3.0.0", "3.1.0", "3.2.0", "3.3.0", "3.3.1", "3.4.0", "3.5.0", "3.6.0", "3.7.0", "3.8.0", "3.9.0", "3.10.0", "3.11.0", "3.12.0", "3.13.0"]
 
         var isCompatible = compatVers.contains(AMDRyzenCPUPowerManagementVersion)
         if !isCompatible {